+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - ✨(backend) make the LiveKit default video codec configurable
+- ✨(fullstack) let an operator forbid public rooms
 
 ### Changed
 
@@ -40,7 +41,6 @@ and this project adheres to
 - 🔊(backend) log request duration in Gunicorn workers
 - 📈(frontend) track missing lobby participant on accept/reject
 - ✨(backend) sort waiting participants by their arrival time
-- ✨(fullstack) let an operator forbid public rooms
 
 ### Changed